License Activation feature
License Activation templates have been added to the Welcome page
templates folder.

For a fresh install or for upgrading users who do not have an active
license key enabled, the plugin will now first ask the admin to activate
the license and then move on to the Welcome page.

Instructions for code changes have been added in the Instructions file.